change to new SCRATCH variable usage on HPCC

+++ pipeline/03_jointGVCF_call_slice.sh
@@ -11,18 +11,23 @@ module load java/13
 module load bcftools
 module load parallel
 module load yq
-
+module load workspace/scratch
 source config.txt
 
-declare -x TEMPDIR=$TEMP/$USER/$$
+#declare -x TEMPDIR=$TEMP/$USER/$$
 
-cleanup() {
+#cleanup() {
 	#echo "rm temp is: $TEMPDIR"
-	rm -rf $TEMPDIR
-}
+#	rm -rf $TEMPDIR
+#}
 
 # Set trap to ensure cleanupis stopped
-trap "cleanup; rm -rf $TEMPDIR; exit" SIGHUP SIGINT SIGTERM EXIT
+#trap "cleanup; rm -rf $TEMPDIR; exit" SIGHUP SIGINT SIGTERM EXIT
+if [ ! -z $SCRATCH ]; then
+	TEMPDIR=$SCRATCH
+else
+	TEMPDIR=$TEMP/$USER/$$
+fi
 
 GVCF_INTERVAL=1
 N=${SLURM_ARRAY_TASK_ID}
@@ -179,6 +184,6 @@ do
 	fi
 done
 
-if [ -d $TEMPDIR ]; then
+if [[ -z $SCRATCH && -d $TEMPDIR ]]; then
 	rmdir $TEMPDIR
 fi
